Make `Retry` public and to its own file
