+++ internal/e2eutil/helper.go
@@ -1,0 +1,34 @@
+package e2eutil
+
+import (
+	"context"
+	"time"
+
+	"github.com/pkg/errors"
+)
+
+var ErrContinueRetry = errors.New("continue Retry")
+
+// Retry retries the given function until reached timeout. The function should
+// return ErrContinueRetry to indicate another retry.
+func Retry(timeout time.Duration, fn func() error) error {
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+
+	for {
+		select {
+		case <-ctx.Done():
+			if ctx.Err() == context.DeadlineExceeded {
+				return errors.Errorf("Retry timed out in %s", timeout)
+			}
+			return ctx.Err()
+		default:
+			err := fn()
+			if err != ErrContinueRetry {
+				return err
+			}
+		}
+
+		time.Sleep(100 * time.Millisecond)
+	}
+}
